added new task2 in TreadExample
